[DOC] Update strong_type documentation

+++ test/snippet/core/detail/strong_type_adding_skills.cpp
@@ -2,16 +2,16 @@
 
 using seqan3::operator|;
 
-struct error : seqan3::detail::strong_type<unsigned, error, seqan3::detail::strong_type_skill::decrement |
-                                                            seqan3::detail::strong_type_skill::increment>
+struct error : seqan3::detail::strong_type<uint8_t, error, seqan3::detail::strong_type_skill::decrement |
+                                                           seqan3::detail::strong_type_skill::increment>
 {
-    using seqan3::detail::strong_type<unsigned, error, seqan3::detail::strong_type_skill::decrement |
-                                                       seqan3::detail::strong_type_skill::increment>::strong_type;
+    using seqan3::detail::strong_type<uint8_t, error, seqan3::detail::strong_type_skill::decrement |
+                                                      seqan3::detail::strong_type_skill::increment>::strong_type;
 };
 
 int main()
 {
-    error e{4};
+    error e{4u};
     --e;
     ++e;
 }
